feat: Add quilt import function `rqi`
This function takes a set of quilt patches and applies them to the
appropriate repositories in a `repo` managed project as if we had done a
series of `git am`s. The function currently only looks for patches in
the `.rad/patches` directory but it should be fairly straightforward to
make it work with arbitrary quilt directories in future.