fix: Don't put upstream in merge message
This worked because I happened to be in a project working tree with the
branch checked out, but we generally can't know what the upstream of a
branch being merged is for all projects. Instead, we'll just put the
name of the branch to be merged in the merge commit message and rely on
the user to add mroe context if they want.