fix: Avoid `repo forall` to fix interactive cmds
`repo` v2.13 dropped support for interactive `forall` commands in
fbab606 [0] and pointed out that they weren't reliable anyway due to the
default `-j2` behaviour. With this in mind, it's easier to just get the
list of repo paths and loop over them ourselves. This does mean we lose
a bunch of the environment variables provided by `forall` but we only
currently use `REPO_PATH` so we just export that ourselves now.